+++ client/src/components/layout/Home.module.css
@@ -43,16 +43,17 @@
 
 .widget3{
   border: 2px solid paleturquoise;
-  height: 37vh;
+  height: 328px;
   border-radius: 25px;
   width: 25vw;
   margin: 1em;
+  overflow: auto;
   
 }
 
 .widget4{
   border: 2px solid paleturquoise;
-  height: 37vh;
+  height: 328px;
   width: 65vw;
   border-radius: 25px;
   margin: 1em;
