+++ client/src/components/layout/WeCare.jsx
@@ -1,9 +1,8 @@
-import Player from "../spotify/Login/Player";
 
 
 const Wecare = () => {
   return(
-    <div><Player></Player></div>
+    <div></div>
   )
 }
