+++ client/src/components/spotify/Login.module.css
@@ -1,4 +1,4 @@
-.login-btn {
+.loginBtn {
   width: 180px;
   padding: 15px 0px;
   text-align: center;
